# apps/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Body_update_profile_multipart_me_profile_patch": {
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "currentPassword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentpassword"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "removeAvatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Removeavatar"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "username"
        ],
        "title": "Body_update_profile_multipart_me_profile_patch",
        "type": "object"
      },
      "Body_update_profile_multipart_v1_me_profile_patch": {
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "currentPassword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentpassword"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "removeAvatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Removeavatar"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "username"
        ],
        "title": "Body_update_profile_multipart_v1_me_profile_patch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Family Recipe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/me/favorites": {
      "get": {
        "operationId": "my_favorites_me_favorites_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Favorites",
        "tags": [
          "me"
        ]
      }
    },
    "/me/password": {
      "put": {
        "operationId": "change_password_me_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password",
        "tags": [
          "me"
        ]
      }
    },
    "/me/profile": {
      "patch": {
        "description": "Update the current user's profile, including optional avatar upload.\n\nMultipart-only (`multipart/form-data`) per parity with\n`src/app/api/me/profile/route.ts`. Field names are flat (not a `payload`\nJSON envelope) to match Next; the migration plan's payload-envelope\nshape is tracked as a follow-up under #188.\n\nBehavior:\n- Validates `name`, `email`, `username` via Pydantic's `UpdateProfileRequest`\n  (same constraints as Next's `updateProfileSchema`).\n- If `email` or `username` differ from the current values, requires a\n  matching `currentPassword` (mirrors Next's `requiresPassword` branch).\n- `avatar` file (\u22645MB, JPEG/PNG/WEBP) writes its storage key to the\n  `avatarStorageKey` column (Postgres column `avatar_url`, mapped via\n  `@map` in the Prisma schema). The DB stores an opaque storage key,\n  not a URL \u2014 the legacy column name is kept for compatibility with\n  Next-era rows that still hold pre-migration values, but the Prisma\n  attribute is canonical.\n- `removeAvatar=true` clears the column.\n- On sensitive-field change (email or username), the session cookie is\n  cleared so the next request re-authenticates \u2014 matches Next's\n  `clearSessionCookie` call.\n- P2002 unique-constraint violations on email/username surface as\n  `409 CONFLICT` per the Next contract.",
        "operationId": "update_profile_multipart_me_profile_patch",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_profile_multipart_me_profile_patch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile Multipart",
        "tags": [
          "me"
        ]
      },
      "put": {
        "description": "Legacy JSON profile-update handler. Kept alive for Phase-2 clients.\n\nThe canonical Phase-3 endpoint is `update_profile_multipart` above\n(`PATCH /v1/me/profile`), which mirrors Next's contract (flat form\nfields, multipart body, optional avatar, sensitive-field-change\npassword check, 409 on duplicate). This PUT remains until Phase-4\ncutover removes the legacy non-`/v1/` routes (#38) \u2014 keep behaviour\nhere drift-free of the PATCH unless a deliberate parity change is\nbeing made on both sides.",
        "operationId": "update_profile_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "me"
        ]
      }
    },
...
    "/profile/favorites": {
      "get": {
        "operationId": "my_favorites_profile_favorites_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Favorites",
        "tags": [
          "profile"
        ]
      }
    },
...
    "/v1/me/favorites": {
      "get": {
        "operationId": "my_favorites_v1_me_favorites_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Favorites",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/password": {
      "put": {
        "operationId": "change_password_v1_me_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/profile": {
      "patch": {
        "description": "Update the current user's profile, including optional avatar upload.\n\nMultipart-only (`multipart/form-data`) per parity with\n`src/app/api/me/profile/route.ts`. Field names are flat (not a `payload`\nJSON envelope) to match Next; the migration plan's payload-envelope\nshape is tracked as a follow-up under #188.\n\nBehavior:\n- Validates `name`, `email`, `username` via Pydantic's `UpdateProfileRequest`\n  (same constraints as Next's `updateProfileSchema`).\n- If `email` or `username` differ from the current values, requires a\n  matching `currentPassword` (mirrors Next's `requiresPassword` branch).\n- `avatar` file (\u22645MB, JPEG/PNG/WEBP) writes its storage key to the\n  `avatarStorageKey` column (Postgres column `avatar_url`, mapped via\n  `@map` in the Prisma schema). The DB stores an opaque storage key,\n  not a URL \u2014 the legacy column name is kept for compatibility with\n  Next-era rows that still hold pre-migration values, but the Prisma\n  attribute is canonical.\n- `removeAvatar=true` clears the column.\n- On sensitive-field change (email or username), the session cookie is\n  cleared so the next request re-authenticates \u2014 matches Next's\n  `clearSessionCookie` call.\n- P2002 unique-constraint violations on email/username surface as\n  `409 CONFLICT` per the Next contract.",
        "operationId": "update_profile_multipart_v1_me_profile_patch",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_profile_multipart_v1_me_profile_patch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile Multipart",
        "tags": [
          "me"
        ]
      },
      "put": {
        "description": "Legacy JSON profile-update handler. Kept alive for Phase-2 clients.\n\nThe canonical Phase-3 endpoint is `update_profile_multipart` above\n(`PATCH /v1/me/profile`), which mirrors Next's contract (flat form\nfields, multipart body, optional avatar, sensitive-field-change\npassword check, 409 on duplicate). This PUT remains until Phase-4\ncutover removes the legacy non-`/v1/` routes (#38) \u2014 keep behaviour\nhere drift-free of the PATCH unless a deliberate parity change is\nbeing made on both sides.",
        "operationId": "update_profile_v1_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "me"
        ]
      }
    },
...
    "/v1/profile/favorites": {
      "get": {
        "operationId": "my_favorites_v1_profile_favorites_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Favorites",
        "tags": [
          "profile"
        ]
      }
    },
...
  }
}